RCTF2022_CatSpy/RCTF_catspy/requirements.txt
flask==2.2.2
Pillow==9.1.1